Move AnnotatedTree to its own file.
