+++ .gitignore
@@ -3,3 +3,5 @@ target/
 .idea/
 *.iml
 *.DS_Store
+*.swp
+tags
